Identifying variables in constraints

JuMPIn.identify_unique_variables — Function

identify_unique_variables(constraints::Vector)::Vector{JuMP.VariableRef}

Return a vector of variables that participate in the provided constraints.

Each variable appears at most one time in the returned vector. If we receive a vector, we just assume it is a vector of constraints. This is because I couldn't get an argument of type Vector{ConstraintRef} to work...

Example

julia> using JuMP
julia> import JuMPIn as ji
julia> m = Model();
julia> @variable(m, v[1:3]);
julia> @constraint(m, eq_1, v[2] == 1);
julia> @NLconstraint(m, eq_2, v[2]*v[3]^1.5 == 2);
julia> vars = ji.identify_unique_variables([eq_1, eq_2]);
julia> display(vars)
2-element Vector{VariableRef}:
 v[2]
 v[3]

identify_unique_variables(model::JuMP.Model, include_inequality::Bool=false)

Return a vector of variables that participate in constraints in the model.

Each variable appears at most one time in the returned vector.

identify_unique_variables(constraint::JuMP.ConstraintRef)

Return a vector containing the variables that participate in this constraint.

Each variable appears at most one time in the returned vector.

identify_unique_variables(
    constraint::JuMP.ConstraintRef,
    index::JuMP.ConstraintIndex,
)::Vector{JuMP.VariableRef}

Return a vector containing the variables that participate in this constraint.

Each variable appears at most one time in the returned vector. The index argument is provided so we know whether we have a "regular" constraint or a nonlinear constraint.

identify_unique_variables(
    constraint::JuMP.ConstraintRef,
    index::MathOptInterface.Nonlinear.ConstraintIndex,
)::Vector{JuMP.VariableRef}

Return a vector containing the variables that participate in this constraint.

Each variable appears at most one time in the returned vector. The index argument is provided so we know whether we have a "regular" constraint or a nonlinear constraint.

identify_unique_variables(fcn)::Vector{JuMP.VariableIndex}

Return the variables that appear in the provided MathOptInterface function.

Implementation

Only ScalarQuadraticFunction and ScalarAffineFunction are supported. This can be changed there is demand for other functions. For each type of supported function, the _get_variable_terms function should be defined. Then, for the type of each term, an additional identify_unique_variables function should be implemented.

identify_unique_variables(term)

Return the variables that participate in the provided term of a MathOptInterface function.

A variable appears at most one time in the returned vector.

Implementation

Currently implemented only for ScalarAffineTerm and ScalarQuadraticTerm.

JuMPIn._get_variable_terms — Function

_get_variable_terms(fcn)

Return a tuple of vectors of terms for the provided MathOptInterface function.

Implementation

Currently implemented only for ScalarQuadraticFunction and ScalarAffineFunction.