Incidence matrix

JuMPIn.incidence_matrix — Function

incidence_matrix(constraints, variables)::SparseMatrixCSC

Return a SparseMatrixCSC with entries corresponding to incidence between the provided constraints and variables.

Rows correspond to constraints, columns correspond to variables, and matrix entries correspond to edges in the incidence graph. All matrix entries have values of 1.0.

incidence_matrix(igraph::IncidenceGraphInterface)::SparseMatrixCSC

Return the incidence matrix associated with the provided IncidenceGraphInterface.