Identifying equality constraints

JuMPIn.get_equality_constraints — Function

get_equality_constraints(model::JuMP.Model)::Vector{JuMP.ConstraintRef}

Return a vector of equality constraints in the provided model.

Example

julia> using JuMP
julia> import JuMPIn as ji
julia> m = Model();
julia> @variable(m, v);
julia> @constraint(m, v == 1);
julia> eq_cons = ji.get_equality_constraints(m);
julia> display(eq_cons)
1-element Vector{ConstraintRef}:
 eq_con_1 : v = 1.0

JuMPIn.is_equality — Function

is_equality(constraint::JuMP.Model)::Bool

Detect whether a constraint is an equality constraint.

JuMPIn.set_implies_equality — Function

set_implies_equality(set::T)::Bool where T<:MathOptInterface.AbstractSet

Detect whether the set defines an equality constraint, i.e. is a singleton.

Implementation

Methods are defined for the following MathOptInterface.Sets:

MathOptInterface.EqualTo

MathOptInterface.Interval

If a MathOptInterface.AbstractVectorSet is provided, an error is raised. For any other type of set, false is returned. To support additional types of constraints in is_equality and get_equality_constraints, additional methods of set_implies_equality should be defined.