Algorithm interfaces

JuMPIn.IncidenceGraphInterface — Type

IncidenceGraphInterface(model; include_inequality = false)

A bipartite incidence graph of JuMP constraints and variables.

This is the primary data type accepted by the algorithms implemented in the remainder of this module. This type can be instantiated with a JuMP model or a tuple of (graph, con_node_map, var_node_map), as returned by get_bipartite_incidence_graph.

Note that the fields of this struct are private, and may change behavior in a future release without warning.

Example

using JuMP
import JuMPIn as ji
m = Model()
@variable(m, v[1:3])
@constraint(m, eq_1, v[1] + v[3]^2 == 1.0)
@NLconstraint(m, eq_2, v[1]*v[2]^1.5 == 2.0)
graph = ji.IncidenceGraphInterface(m)

JuMPIn.get_adjacent — Function

get_adjacent(
    igraph::IncidenceGraphInterface,
    constraint::JuMP.ConstriantRef,
)::Vector{JuMP.VariableRef}

Return the variables adjacent to a constraint in an incidence graph.

get_adjacent(
    igraph::IncidenceGraphInterface,
    variable::JuMP.VariableRef,
)::Vector{JuMP.ConstraintRef}

Return the constraints adjacent to a variable in an incidence graph.

Example

julia> using JuMP
julia> import JuMPIn as ji
julia> m = Model();
julia> @variable(m, v[1:3]);
julia> @constraint(m, eq_1, v[1] + v[3] == 1);
julia> @NLconstraint(m, eq_2, v[1]*v[2]^3 == 2);
julia> igraph = ji.IncidenceGraphInterface(m);
julia> adj_cons = ji.get_adjacent(igraph, v[1]);
julia> display(adj_cons)
2-element Vector{ConstraintRef}:
 eq_1 : v[1] + v[3] = 1.0
 v[1] * v[2] ^ 3.0 - 2.0 = 0

JuMPIn.maximum_matching — Function

maximum_matching(igraph::IncidenceGraphInterface)::Dict

Compute a maximum matching of variables and constraints in the incidence graph.

The returned Dict maps JuMP ConstraintRefs to their matched VariableRefs.

Example

julia> using JuMP
julia> import JuMPIn as ji
julia> m = Model();
julia> @variable(m, v[1:3]);
julia> @constraint(m, eq_1, v[1] + v[3] == 1);
julia> @NLconstraint(m, eq_2, v[1]*v[2]^3 == 2);
julia> igraph = ji.IncidenceGraphInterface(m);
julia> matching = ji.maximum_matching(igraph);
julia> display(matching)
Dict{ConstraintRef, VariableRef} with 2 entries:
  v[1] * v[2] ^ 3.0 - 2.0 = 0 => v[2]
  eq_1 : v[1] + v[3] = 1.0 => v[1]

JuMPIn.dulmage_mendelsohn — Function

dulmage_mendelsohn(igraph::IncidenceGraphInterface)

Return the Dulmage-Mendelsohn partition of variables and constraints in an incidence graph.

The returned type is a Tuple of two NamedTuples, (con_dmp, var_dmp). These NamedTuples have the following fields:

con_dmp:

underconstrained – The constraints matched with variables that can possibly be unmatched in a maximum cardinality matching

square – The constraints that cannot possibly be unmatched in a maximum matching

overconstrained – The constraints that are matched, but can possibly be unmatched in a maximum matching

unmatched – The constraints that are not matched in the maximum matching that was found

var_dmp:

unmatched – The variables that are not matched in the maximum matching that was found

underconstrained – The variables that can possibly be unmatched in a maximum matching

square – The variables that cannot possibly be unmatched in a maximum matching

overconstrained – The variables matched with constraints that can possibly be unmatched in a maximum cardinality matching

The Dulmage-Mendelsohn partition groups nodes in a bipartite graph into three unique subsets. In the application to constraints and variables, these may be thought of as:

the "overconstrained subsystem", which has more constraints than variables,

the "underconstrained subsystem", which has more variables than constraints,

and the "square subsystem", which has the same number of variables as constraints

In the NamedTuples returned by this function, the constraints in the overconstrained subsystem are split into overconstrained and unmatched, while the variables in the underconstrained subsystem are split into underconstrained and unmatched. This is because it is useful to explicitly check whether there are any unmatched variables and constraints, and also useful to recover the maximum matching by zip-ing corresponding variables and constraints.

Example

julia> using JuMP
julia> import JuMPIn as ji
julia> m = Model();
julia> @variable(m, v[1:4]);
julia> @constraint(m, eq_1, v[1] + v[3] == 1);
julia> @NLconstraint(m, eq_2, v[1]*v[2]^3 == 2);
julia> @constraint(m, eq_3, v[4]^2 == 3);
julia> igraph = ji.IncidenceGraphInterface(m);
julia> con_dmp, var_dmp = ji.dulmage_mendelsohn(igraph);
julia> # Assert that there are no unmatched constraints
julia> @assert isempty(con_dmp.unmatched);
julia> display(var_dmp.unmatched)
1-element Vector{VariableRef}:
 v[3]
julia> display(var_dmp.underconstrained)
2-element Vector{VariableRef}:
 v[1]
 v[2]
julia> display(con_dmp.underconstrained)
2-element Vector{ConstraintRef}:
 eq_1 : v[1] + v[3] = 1.0
 v[1] * v[2] ^ 3.0 - 2.0 = 0
julia> display(var_dmp.square)
1-element Vector{VariableRef}:
 v[4]
julia> display(con_dmp.square)
1-element Vector{ConstraintRef}:
 eq_3 : v[4]² = 3.0
julia> # As there are no unmatched constraints, the overconstrained subsystem
julia> # is empty.